Allow to change toolbar font separately again
commit 86be8d49d8988fb6c35e847ed9e0aad3e8514208
Author: Henrik Fehlauer <rkflx@lab12.net>
Date:   Fri Jul 14 09:04:25 2017 +0200

Allow to change toolbar font separately again

Summary:
The fonts KCM supports changing the toolbar font, but with the port to
QPA/Plasma 5 in aa16577f this stopped working and toolbars just used the
general font. This is fixed by setting QPlatformTheme's ToolButtonFont
to the toolbar font chosen. However, there are some caveats:

Default toolbar font size is 9, while the general font size is 10.
This would result in a sudden shrink in toolbar widths everywhere (or
        worse for weird leftover settings users might have tried unsuccessfully).

In some places QToolButtons are used outside of toolbars, sometimes
right next to QPushButtons. As due to the bug both had the same font,
      this might now lead to unexpected differences.

      Therefore, to keep visual consistency with past Plasma 5 releases, let's
      change the default toolbar font size to 10 and provide a kconf_update
      script to copy the general font to the toolbar font once.

      BUG: 358254
      FIXED-IN: 5.11.0

      Test Plan:
      make install, run kconf_update: toolBarFont in kdeglobals gets updated
      immediately

      Change "Toolbar" font in "kcmshell5 fonts", start your favorite KDE
      and Qt-only applications: toolbar font and/or font size changed
      accordingly

      Reviewers: #plasma, davidedmundson, fvogt

      Reviewed By: #plasma, davidedmundson, fvogt

      Subscribers: fvogt, davidedmundson, jensreuterberg, plasma-devel

      Tags: #plasma

      Differential Revision: https://phabricator.kde.org/D6697